Prepare repository for application
The fluid, geodesic, geometry, tests, and benchmark folders have been deleted since they will not be used for this application. Also, a results folder was generated to dump all the simulation, analysis, and plotting results. This folder will remain empty in the public repository. A configure.sh script was also created to expedite configuration with autotools.