Upload solar wind plotting code
Upload a program to output frames (in .silo format) of some solar wind quantities over the first quadrant (spherical symmetry) of a meridional slice while Voyager 2 was in the heliosheath. The information produced should be enough to visualize the current sheet folds, termination shock, and heliopause, as well as the unipolar/sectored regions delimited in the simulation. The Voyager 2 trajectory is also outputed (is CSV format).