Major update
The numerical derivative routines for the drifts and divergence of the diffusion tensor have been fixed to only use the gyroradius, not the minimum of the gyroradius and dmax. Also, simulation codes for hydrogen have been generated and all plotting scripts modified to allow for this specie. Finally, the README.md file was updated to be more descriptive and useful, but it still needs some more work before publication.